[Promise] Integrate with Asyncify and JSPI
Add an `emscripten_promise_await_sync` function that appears to synchronously
wait for a promise to be resolved using either Asyncify or JSPI. This allows
users to implement an async-sync boundary wherever they want in their code
rather than just at the Wasm-JS boundary. Another convenience is that
synchronously waiting for a promise makes it safe for that promise to access
stack data that is live at the time of the wait.